Call BQ:set_queue_version when policy is removed
When the queue args don't include the version and a v2 policy
is removed, we will get an 'undefined' value. In that case we
still want to call BQ:set_queue_version because we have to go
back to the default version v1.

(cherry picked from commit f35ca2a19b245d192f89755b424aee9084afcc16)