Fortify v1->v2 conversion resume
Don't write messages if they are already in the v2 index. This
can happen when the node crashes after writing to v2 but before
deleting the v1 files, but only when the per-queue store syncs
in the middle of writing a segment (because otherwise the messages
are dropped from the v2 index on recovery and re-written again
as if they were never there).

(cherry picked from commit 8af2ede58bf61921b40da9197041048b4cbc28d4)