CQ property suite: Add vhost restart
This helps us test cases where the queue restarts cleanly.
Because it is not completely deterministic, there is some
clever handling of messages to accept messages that were
acked by the client but the server didn't ack before the
restart, and messages published without confirms that
never made it to the server before it restarted.

There is potential to improve confirms handling for that
scenario but that is left as an exercise for a later time.