Set a default cluster name when the cluster name is not configured.
If the cluster name is not set, 'rabbit_reader' will get the host name when filling out the cluster name into the server properties when creating a new link.

In some cases, 'inet:gethostbyname' is time-consuming. See my testing log:

2021-12-31 03:02:32.613676+08:00 [warn] <0.897.0> receive amqp HEADER
2021-12-31 03:02:32.613845+08:00 [warn] <0.897.0> register_connection return
2021-12-31 03:02:32.613880+08:00 [warn] <0.897.0> before  inet:gethostname
2021-12-31 03:02:32.613988+08:00 [warn] <0.897.0> return  inet:gethostname ,Hostname:"paas-controller1"
2021-12-31 03:02:36.127212+08:00 [warn] <0.897.0> return  inet:gethostbyname ,Res:"paas-controller1"
2021-12-31 03:02:36.128272+08:00 [warn] <0.897.0> server_properties before return

As can be seen from the log, 'inet:gethostbyname' takes more than 3 seconds. 
that is: 2021-12-31 03:02:32.613988+08:000 ~~ 2021-12-31 03:02:36.127212+08:00