CQ v2: Fix infinite loop in convert_to_lazy in rare scenarios
In some rare scenarios (involving multiple segment files
with few messages each, and many clean/dirty restarts),
it would be possible for a message to be in q2 and never
pushed to disk. By setting the io_batch_size to 0 for this
specific operation we ensure that ALL messages get pushed
to disk and nothing remains afterwards (up to 4096 could
remain before both for v1 and v2).