Fix performance regression when stream_queue feature is not enabled
If the quorum_queue and stream_queue features were not enabled the
classic queues would use the old message format for confirms.

This would cause the channel to scan its entire queue type state
map to find the queue name corresponding to the pid. Particularly
when using large ratio fanouts (1:600) this could cause very poor
performance.

Luckily we use the queue type's monitor_registery to get exactly the
information we need quickly which is what this change implements.