fix shovels crashing with delete-after=0
When a shovel was added with delete-after=0 it crashed immediatelly,
causing to never be cleaned up. This caused RabbitMQ to not boot if
the shovel parameter existed.

The fix makes sure the validation is correct on this value, and when
a shovel starts it will let it start with a 0 delete-after.

(cherry picked from commit f934fde55a20d750ac96db897d551ecefa78ed14)
(cherry picked from commit e9eed6c8406c1a3a5f0753e5bce38de10ef3fba8)