rabbit_fifo: fix SAC bug when a cancelled consumer is revived
If a cancelled single active consumer with a message checked out
is revived through another checkout commend
we'd previously would have overwritten the active consumer
and lost track of the in-flight message. This change handles
this case and merges the cancelled and new consumer with the
same identity.

(cherry picked from commit c55fd217adeb195517f1657867382309764ccf52)