Make sure outdated pending acks are settled
as in, acknowledged to the queue replica.

Previous iteration in #4004 only kept track of stale delivery tags
(they were pending acknowledged when queue leader failed),
which meant "unknown" (stale) basic.ack operations were handled
gracefully but no actual settlement with the queue process took place.

With this change, we keep track of #pending_ack records and not just
delivery tags, which retains all context necessary for settlement,
namely the name of the queue and message ID.