CQv2 index: Increase threshold before flushing to disk
Before it was flushing when there was 100 entries which was way
too low, especially since entries do not include the message
content. Change it to flushing when there's a segment file
worth of entries. Also change updates to flush them when
there's half a segment file worth of updates (which in
practice doesn't change much from before).

Huge shootout to Michal Kuratczyk for all the testing and
bugging me long enough to make me understand something was
wrong.