Make Osiris listener registration idempotent
Several listeners for the same offset can be registered,
which causes problems for low rate, where consumers hit
the end of the stream often. Then Osiris messages accumulate
in the process mailbox, which slows it down considerably after
some time.

This commit makes the listener registration idempotent for a given
offset.