Support quorum queue leader locator
Prior to this commit:
1. When a new quorum queue was created, the local node + random nodes
   were selected as replicas.
2. Always local node became leader.

For example, when an AMQP client connects to a single RabbitMQ node and
creates N quorum queues, all N leaders will be on that node and replicas
are not evenly distributed across the RabbitMQ cluster.
If N is small and the RabbitMQ cluster has many nodes, some nodes might
not host any quorum queue replicas at all.

After this commit:
1. When a new quorum queue is created, the local node + RabbitMQ nodes
   with least quorum queue replicas are selected.
   This will nicely distribute the quorum queue replicas across the
   RabbitMQ cluster.
2. Support (x-)queue-leader-locator argument / policy with
    * client-local (stays the default)
    * random
    * least-leaders
    The same settings are already available for streams.