Speed up re-importing queue definitions
The channel does not declare a queue that already exists.
Let's do the same when importing definitions.

On a 5-node cluster re-importing 40,000 classic non-mirrored queues with
queue-master-locator set to min-masters takes 11min32sec before this
change and 3sec after this change. So it's 230 times faster after this
change.

With more queues, the speed-up factor will be much greater.