Do not declare queues that already exist
when importing definitions.

The channel is bailing out early when the client declares a queue that
already exist.
Let's do the same for import definitions.

This considerably speeds up the use case where thousands of already
declared quorum queues get re-imported with queue-leader-locator set
to least-leaders or with an replication factor smaller than the
RabbitMQ cluster size.