Avoid expensive least replicas selection
If there are a lot of queues, fetching them from the database
and counting the number of replicas per node will be costly.

Random replica selection is good enough, will balance the RabbitMQ
cluster in the long term and this was the only way to select replicas
before this PR.

The only downside of random replica selection is when there are many
RabbitMQ nodes in the cluster, say 7, but only few quorum queues are
created, the randomness might not place replicas on all 7 nodes.

Therefore, in this commit we ensure even distribution of replicas across
RabbitMQ nodes only if the total number of queues is not high
(hard-coded to 1000 right now).

Only the total number of queues can be cheaply counted by calling
```
mnesia:table_info(rabbit_queue, size)
```
Number of queues by queue type or replicas per node cannot be cheaply
counted as this would require a full table scan.
Note that ets:select_count/2 would also need to scan the whole table.

There can be edge cases such as the following:
5 node RabbitMQ cluster with 10k non-HA queues on node 1, 2, and 3
but no queues on node 4, and 5.
New 3-replica quroum queue gets declared and the random replica selection
could place the replicas on node 1, 2, and 3.

The benefit of avoiding expensive least replica selection greatly outweighs
the downside of such edge cases.