CQ: Fix potential crash when collecting stats
The CQ is looking up itself when collecting CMQ stats.
There seems to be a potential race condition on init
where the initial stats update happens before the
queue can be looked up. To prevent the crash, we
ensure that when the queue is not found we don't
return stats for these items.