Validate queue arguments
Throw a PRECONDITION_FAILED error when a queue is created with an
invalid argument.
There can be RabbitMQ plugins and extensions out there defining arbitrary
queue arguments. Therefore we only check that a queue is not declared
with arguments that we know are supported only by OTHER queue types.

The benefit of this change is that queues are not mistakenly declared
with the wrong arguments. For example, a stream should not be
declared with `x-dead-letter-exchange` because this argument is not
supported in streams. It is only supported by classic and quorum queues.
Instead of silently allowing this argument and users wondering why the stream
does not dead letter any messages, it's better to fail early with an
meaningful error message.

We still allow any other arbitrary queue arguments.
(Therefore and unfortunately, when `x-dead-letter-exchange` is misspelled,
it will still be accepted as an argument).

For argument equivalence, we only validate the arguments that are
supported for a queue type. There is no benefit in validating any other
arguments that are not supported anyways.

(cherry picked from commit ef3c9cd526ea0a13c8ea808ba5581318cacb7dfa)