Correct stream queue capabilities
Delete arguments:

x-dead-letter-exchange
x-dead-letter-routing-key
x-max-length
x-single-active-consumer

None of these arguments are supported in streams.

Dead lettering does not apply to streams.
x-max-length does not apply, only x-max-length-bytes and x-max-age apply.

Single active consumer does not work on the queue level, it will only
work on the consumer level. This is different to single active consumer
for classic and quorum queues.