improve credit handling and header conversion
The AMQP 1.0 plugin grants publishers 65000 credit to the senders by default.
This PR makes the credit given configurable.

The AMQP 1.0 plugin grants credit when the message is received, not when it is
placed into the queue. This means that an AMQP 1.0 sender who does not care
about settlement can overload the receiver process leading to high memory
usage.

The AMQP 1.0 plugin creates a classic queue by default. This is annoying
because the client can not ensure that the target exists (it needs to use
/amq/queue/Q target instead of the name of the queue.
Now the plugin can be configured to use Quorum Queues.

Sending an unsettled message on a link set as 'settled' is now a protocol
error. Before this change it was leading to crashes of the AMQP process with a
crash of the link.

Implements optional conversion of AMQP 1.0 headers to 0.9.1 headers for some
headers types (utf8, uint, bool). Implements #2591

Implements optional conversion of AMQP 0.0.1 headers to AMQP 1.0 App Properties.

Also fixes a bug in the amqp10_client where the client can run out of credit
and get stuck because of bad accounting.