Return empty array when JMS topic exchange state does not exist
Instead of exiting. A few messages can get routed to a deleted
exchange. The exchange has no state in Mnesia, so the previous
code would return an error that would bubble up to the publisher
and closing its channel. The exchange is created for a subscriber,
so this would mean the deletion of a subscriber would impact
a publisher.

This commit makes the exchange return an empty array if no state is found.
This is less drastic and hopefully more appropriate than returning an error.
The exchange state should always be found under normal conditions.

See https://github.com/rabbitmq/rabbitmq-server/discussions/4616