rabbit_classic_queue_index_v2: Support copy-on-write filesystems
Copy-on-write filesystems don't support file preallocation by design. In
this case, posix_fallocate(2) and friends will fail. For example it
fails with EINVAL on FreeBSD+ZFS.

There is no point in filling the file with zeroes to mimic
preallocation, like it is done for OSes/filesystems which don't support
this feature. However, we still do it anyway because the rest of the
code assumes that the file was preallocated.

Fixes #4720.