Save memory when acking messages to quorum queue
List `MsgIds` has fewer elements than list `Settles`.
Therefore, put it on the left side of the `++` operator.

The memory flame graph revealed that before this commit
5%-8% of all mmap() system calls happened in function
rabbit_fifo_client:settle/3.

After this commit only 1.6% - 1.8% of all mmap() system calls happen in
this function.

Note that we cannot do the same for discarded messages (`Discards`)
because the order in which messages will be dead lettered need to be
preserved.