rabbit_feature_flags: Support required feature flags
Required feature flags are existing feature flags which were introduced
some time ago and we want to make them required and remove the old
compatibility code.

Required feature flags MUST be enabled before upgrading to the RabbitMQ
version where they are marked required. If they are not, RabbitMQ will
refuse to start. The reason is their corresponding compatibility and
migration code is gone.

To make a feature flag required, its `stability` property needs to be
set to `required`.

V2: Use the word `required` instead of `mandatory` in the initial patch,
    as discussed with @michaelklishin.