feature_flags_SUITE: Test required feature flags using `quorum_queue`
Our test framework can inject feature flags, but we would need a special
handling for required injected feature flags which would not test the
regular code path.

Therefore we rely on the `quorum_queue` feature flag, now required, to
test that it is correctly enabled on boot and when clustering.