feature_flags_SUITE: Use `stream_queue` instead of `quorum_queue` in testcases
`quorum_queue` is now required and can't be used in tests.

Also, part of the `enable_feature_flag_when_ff_file_is_unwritable`
testcase was commentted out because it relied on the `is_enabled` thing
which was dropped in `rabbit_ff_controller`. This should be introduced
at some point with a more robust design.