Log when acquiring or releasing a global lock
to easily detect dead locks.

Prior to this commit, one caller of function
rabbit_ff_registry_factory:acquire_state_change_lock/0
was logging that it acquires the lock.
However, there are other callers which did not log.
Therefore, it was impossible to detect any dead locks by just
looking at the the logs.