CQ: Fix two accidental state overwrites
The main issue was the one in fetch_by_predicate which resulted
in a double file:close on a single FD, the second failing and
triggering a badmatch. It only occurs for CQv2 in lazy mode
and with a high enough consume rate. It is possible that in
other conditions problems did occur but were not detected.

The other was detected while looking for the original issue,
but might not have created any bug.