Configure default queue type by vhost
This allows operators to override the default queue type on a per-vhost
basis by setting the default_queue_type metadata key on the vhost record.

When a queue is declared without specifiying a queue type (x-queue-type)
and there is a default set for the vhost we will augment the declare arguments
with the default. This allows future declares with and without the x-queue-type
argument to succeed.

Also only change the default _if_ the queue is durable and not
exclusive.