Be more defensive when looking up source exchange
A source exchange should always exist for a given binding.
However, be more defensive in case it does not exist.
In this case, we want to delete the binding from the
rabbit_index_route table as well (if it is even present).
The only reason we do not always delete from the rabbit_index_route
table is to reduce the number of lock conflicts on the same table key.