Bazel: Set $RABBITMQ_FEATURE_FLAGS to the list of required flags in 3.11.0
It was set to the empty string before to make sure no feature flags were
enabled out-of-the-box when starting a new RabbitMQ node.

With 3.11.0, several of them will become required and their
corresponding compatibility and migration code goes away. Therefore
older nodes must have them enabled in order to cluster successfully.

This change is only useful in 3.11.x. It must not be backported to older
release branches.