rabbit_mnesia: Sync feature flags before we wait for tables
We want to synchronize feature flags first before we wait for tables
(which is needed to ensure the local view of the tables matches the rest
of the cluster). The reason is that some feature flags may add or remove
tables. In this case the list of tables returned by
`rabbit_table:definitions()` usually depends on the state of feature
flags but this state is local.

For instance, a feature flag may remove a table (so it's gone from the
cluster). If we were to wait for that table locally before synchronizing
feature flags, we would wait forever; indeed the feature flag being
disabled before sync, `rabbit_table:definitions()` would return the old
table.

This fixes the `peer_discovery_class_config_SUITE` testsuite with the
new `listener_records_in_ets` feature flag. Before this patch, we had
the following situation:

* Node 3 acquires the peer discovery lock but fails to cluster with
  Nodes 1 and 2 because they are not ready. It continues as a standalone
  node and enable `listener_records_in_ets` (which deletes the
  `rabbit_listener` table, not replicated at this stage).
* Node 1 is next to acquire the peer discovery lock. It fails to cluster
  with Node 2, but succeeds with Node 3. Unfortunately, it waits for
  tables before synchronizing feature flags. This causes the hang
  because `rabbit_listener` won't appear anytime soon, as it was deleted
  on Node 3.
* Node 2 is spinning, trying to acquire the peer discovery lock.